feat: adding ec2linux ps1 file
