+++ addons/addon-base-raas-appstream/packages/base-raas-appstream-services/lib/plugins/env-sc-connection-url-plugin.js
@@ -39,7 +39,7 @@ async function createConnectionUrl({ envId, connection }, { requestContext, cont
     });
     appStreamUrl = await appStreamScService.getStreamingUrl(requestContext, {
       environmentId: envId,
-      applicationId: 'Notepad',
+      applicationId: 'EC2Linux',
     });
   } else if (connection.scheme === 'rdp') {
     log.debug({
