+++ scripts/app-stream/SETUP.md
@@ -30,6 +30,7 @@ cd ~\Documents
 
 # Pull the Image Builder script from Github
 Invoke-WebRequest -Uri https://raw.githubusercontent.com/awslabs/service-workbench-on-aws/feat-secure-workspace-egress/scripts/app-stream/buildImage.ps1 -OutFile buildImage.ps1
+Invoke-WebRequest -Uri https://raw.githubusercontent.com/awslabs/service-workbench-on-aws/feat-secure-workspace-egress/scripts/app-stream/ec2linux.ps1 -OutFile \`"C:\App\ec2linux.ps1\`"
 
 # Execute Image builder script
 .\buildImage.ps1
