+++ scripts/app-stream/ec2linux.ps1
@@ -1,0 +1,3 @@
+Start-Process -FilePath "C:\Program Files\PuTTY\putty.exe"
+Start-Process -FilePath "C:\Program Files\PuTTY\puttygen.exe"
+Start-Process -FilePath "C:\Windows\System32\notepad.exe"
